scripts/requirements.txt
pytest>=5.3.5
numpy>=1.18.1
nengo>=3.0.0
nengo-dl
matplotlib>=3.1.3
tensorflow>=2.1.0
numba>=0.50.1
setuptools>=45.1.0
tqdm>=4.48.0
pyaml
